feat: makefile to download, filter and deploy one instance per country
It downloads the Natural Earth data, then converts it to GeoJSON, filters it based on a list of ISO
country codes, and deploy the JSON files as well as the index.html and map_all.js files to a
country-specific directory.

The code has been moved to src/ and adapted:
- code style
- load the data
- MapBox key
- remove the increasing difficulty mechanism (show all the country places)

BREAKING CHANGE: The original data is no more available.